# Get rid of custom bufio package.