--- a/.travis.yml
+++ b/.travis.yml
@@ -9,7 +9,6 @@ go:
   - tip
 
 install:
-  - go get gopkg.in/bufio.v1
   - go get gopkg.in/bsm/ratelimit.v1
   - go get github.com/onsi/ginkgo
   - go get github.com/onsi/gomega