--- a/conn.go
+++ b/conn.go
@@ -1,11 +1,12 @@
 package redis
 
 import (
+	"bufio"
 	"net"
 	"time"
-
-	"gopkg.in/bufio.v1"
 )
+
+const defaultBufSize = 4096
 
 var (
 	zeroTime = time.Time{}
@@ -30,7 +31,7 @@ func newConnDialer(opt *Options) func() (*conn, error) {
 		}
 		cn := &conn{
 			netcn: netcn,
-			buf:   make([]byte, 0, 64),
+			buf:   make([]byte, defaultBufSize),
 		}
 		cn.rd = bufio.NewReader(cn)
 		return cn, cn.init(opt)
@@ -101,4 +102,17 @@ func (cn *conn) RemoteAddr() net.Addr {
 
 func (cn *conn) Close() error {
 	return cn.netcn.Close()
+}
+
+func isSameSlice(s1, s2 []byte) bool {
+	return len(s1) > 0 && len(s2) > 0 && &s1[0] == &s2[0]
+}
+
+func (cn *conn) copyBuf(b []byte) []byte {
+	if isSameSlice(b, cn.buf) {
+		new := make([]byte, len(b))
+		copy(new, b)
+		return new
+	}
+	return b
 }