--- a/multi.go
+++ b/multi.go
@@ -115,14 +115,14 @@ func (c *Multi) execCmds(cn *conn, cmds []Cmder) error {
 
 	// Parse queued replies.
 	for i := 0; i < cmdsLen; i++ {
-		if err := statusCmd.parseReply(cn.rd); err != nil {
+		if err := statusCmd.parseReply(cn); err != nil {
 			setCmdsErr(cmds[1:len(cmds)-1], err)
 			return err
 		}
 	}
 
 	// Parse number of replies.
-	line, err := readLine(cn.rd)
+	line, err := readLine(cn)
 	if err != nil {
 		setCmdsErr(cmds[1:len(cmds)-1], err)
 		return err
@@ -143,7 +143,7 @@ func (c *Multi) execCmds(cn *conn, cmds []Cmder) error {
 	// Loop starts from 1 to omit MULTI cmd.
 	for i := 1; i < cmdsLen; i++ {
 		cmd := cmds[i]
-		if err := cmd.parseReply(cn.rd); err != nil {
+		if err := cmd.parseReply(cn); err != nil {
 			if firstCmdErr == nil {
 				firstCmdErr = err
 			}