@vmihailenco 2015-09-03
I can't explain slowdown of `BenchmarkParseReplyString`...