@vmihailenco 2015-09-03
On the 2nd though it must be the cost of `copy`.