@dim 2015-09-03
LGTM in general, but I also think that there is too much going on. A connection has a simple buffer which is filled and drained, seems like the perfect use case for `bytes.Buffer`